Update - Print Composer Added to existing tutorial
This tutorial ended at step 9.  I added 8 more steps specific for creating a Print Composer component.  I also updated some of the existing 9 steps regarding clarity.  Now, a student working on this tutorial will not only be able to geo-reference a historical map using modern-day data, but will also be able to create an image or PDF document that may be submitted for review.